+++ lib/json_cache.dart
@@ -10,6 +10,6 @@ export 'src/json_cache_hive.dart';
 export 'src/json_cache_hollow.dart';
 export 'src/json_cache_local_storage.dart';
 export 'src/json_cache_mem.dart';
-export 'src/json_cache_prefs.dart';
+export 'src/json_cache_shared_preferences.dart';
 export 'src/json_cache_try.dart';
 export 'src/json_cache_wrap.dart';